Misc#signURL - CDN URLSignature
Basic integration of the URL Signature API request into TouchGuild.

Note that TouchGuild doesn't provide a built-in handler for CDN assets. This can be done with a 3rd party library, that is whether existing or needs to be created, or implementing your own way to save CDN assets.

We do not recommend to use the Guilded CDN in its current state, as it complexifies the simple and efficient infrastructure that was needed for a Guilded Application, making it complex (as you can't save in cloud servers while your app is running, that requires the use of an additional structure, making it harder for small, indie projects to make commercial, or even free applications if not stored locally on a computer).